% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylToRec.R
\name{cylToRec}
\alias{cylToRec}
\title{cylToRec}
\usage{
converting coordinates
}
\arguments{
\item{cylindrical}{radius, polar angle, z}
}
\value{
x, y, z tuple
}
\description{
function converts from theta, radius, z cylindrical coordinates to x, y, z coordinates
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
