% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylToSphere.R
\name{cylToSphere}
\alias{cylToSphere}
\title{cylToSphere}
\usage{
converting coordinates
}
\arguments{
\item{cylindrical}{radius, polar angle, z}
}
\value{
x, y, z tuple
}
\description{
converts from cylindrical coordinates to spherical
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
