% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\title{Magnitude}
\usage{
magnitude(x1, y1, z1 = 0, x2 = 0, y2 = 0, z2 = 0)
}
\arguments{
\item{ints}{x1, x2, y1, y2, z1, z2}
}
\value{
numeric representing the magnitude of line segment to point or between points
}
\description{
Finds the size of a line segment or vector
}
\keyword{magnitude}
