% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalizing angles via clock division to reduce angles to one anti-clockwise (clockstupid?) rotations
}
\value{
(0, 2pi] angle
}
\description{
ensures that angles are between 0 and 2pi
}
\examples{
{ normalize(10)}
}
\keyword{angle}
\keyword{normal}
