% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarToRec.R
\name{polarToRec}
\alias{polarToRec}
\title{polarToRec}
\usage{
converting coordinates
}
\arguments{
\item{cylindrical}{radius, polar angle}
}
\value{
x, y tuple
}
\description{
converts polar coordinates to cartesian coordinates
}
\examples{
\dontrun
{ polarToRec( (2 * pi)/3 , 4 )}
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
