% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recToCyl.R
\name{recToCyl}
\alias{recToCyl}
\title{recToCyl}
\usage{
converting coordinates
}
\arguments{
\item{x, }{y, z}
}
\value{
returns theta, cylindrical radius, and z tuple
}
\description{
converts from cartesian coordinates to cylindrical
}
\examples{
\dontrun
{ recToCyl(1,2,3)}
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
