% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recToPolar.R
\name{recToPolar}
\alias{recToPolar}
\title{recToPolar}
\usage{
converting coordinates
}
\arguments{
\item{x, }{y}
}
\value{
cylindrical radius, polar angle
}
\description{
converts rectangular coordinates to polar coordinates
}
\examples{
\dontrun
{ recToPolar(3,4)}
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{remapping,}
