% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recToSphere.R
\name{recToSphere}
\alias{recToSphere}
\title{recToSphere}
\usage{
converting coordinates
}
\arguments{
\item{theta, }{radius, z}
}
\value{
spherical-radius, polar-angle, azimuthal-angle tuple
}
\description{
function converts from x, y, z coordinates to radius, polar theta, azimuthal
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
