% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Slope}
\usage{
finding slope
}
\value{
decimal or integer representing the slope
}
\description{
finds slope of a line with one or two x, y points (3d seems too much)
}
\keyword{slope}
