% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphereToCyl.R
\name{sphereToCyl}
\alias{sphereToCyl}
\title{sphereToCyl}
\usage{
converting coordinates, ass-pains
}
\arguments{
\item{spherical}{radius, polar angle, azimuthal angle}
}
\value{
cylindrical radius, polar angle, z tuple
}
\description{
converts from spherical coordinates to cylindrical
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
