% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphereToRec.R
\name{sphereToRec}
\alias{sphereToRec}
\title{sphereToRec}
\usage{
converting coordinates
}
\arguments{
\item{radius, }{polar angle, azimuthal angle}
}
\value{
x, y, z tuple
}
\description{
converts from cylindrical coordinates to cartesian
}
\keyword{cartesian,}
\keyword{coordinate,}
\keyword{polar,}
\keyword{radians}
\keyword{remapping,}
